==> analysis/5_riverdist.Rmd <==
---
title: "PWS hydraulic distance"
author: "Chase Jalbert & Kyle Shedd"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---

# Introduction

The notebook calculates the river distance of each fish, from the river mouth. It relies on a river network shapefile and 'riverdist' R package. Riverdist was created by ADFG biologist Matt Tyers for tracking fish movements through a river network. More details and vignette can be found here:\
<https://cran.r-project.org/web/packages/riverdist/vignettes/riverdist_vignette.html#Importing%20a%20river%20network>

The river network shapefiles are generally taken from the National Hydrography Database (NHD), however, in somce cases the polylines are not accurate or do not exist. Rivers not appearing in the NHD were delineated, in ArcMap, by simply creating polylines through the fish waypoints (We'll assume these datapoints represent fish in the stream...) and verified using satellite imagery. The shapefiles for Stockdale Creek were modified from the NHD. The shapefiles for Hogan Creek were manually delineated.

High tide marks were provided by Prince William Sound Science Center field crews through Pete Rand. The final distance upstream will be calculated as 1) positive if above high-tide, and 2) negative if below the high-tide mark. The tidal marks are snapped to the streamlines using arcmap and exported as coordinates.

# Setup script

```{r setup}
while(!require(pacman)){ install.packages("pacman") }
  p_load(riverdist, rgdal, sp, tidyverse, lubridate)  # use pacman to load or install + load necessary packages
```

Options for stream are limited to:

    hogan
    stockdale

**NOTE** you will need to run this notebook twice, once for each stream 😉

```{r inputs}
stream = "stockdale" # Enter stream name here

# CSV containing fish data, specifically, Lat/Longs
stream_data = "../data/GIS/inputs/stream_specimens_hogan_stockdale_2013_2016.csv"

# CSV containing high tide points for each stream
tide_data = "../data/GIS/inputs/tide_points.csv"

# Where do you want output to go?
output_path <- paste0("../data/GIS/output/", stream, "_distances.csv")
```

# Setting map projection

**CAUTION** - don't change this unless you want a different projection (and know what you're doing). See this [link](https://mgimond.github.io/Spatial/coordinate-systems-in-r.html) for learning about coordinate systems in R and the definitions below. Briefly:

    +lat_0     Latitude of origin
    +lat_1     Latitude of first standard parallel
    +lat_2     Latitude of second standard parallel
    +lon_0     Central meridian

For re-projecting in Alaska Albers Equal Area projection:

```{r projection}
AKalbers <- "+proj=aea +lat_1=55 +lat_2=65 +lat_0=50 +lon_0=-154
    +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
```

Shapefile location for stream:

```{r prep}
# The shapefile (*.shp) of river network
shapefile_path = normalizePath("../data/GIS/inputs/shapefiles/")
shapefile_name = paste0(stream, "_stream_clean")
```

# Begin Script

First, get the shapefiles and GPS points ready for 'riverdist', then perform basic QC and visualization steps.

Note: Assumptions will be made!

    GPS points do not line up with streams so are snapped to the nearest (straight-line) stream segment
    Sometimes, points in the intertidal extend past stream network, so are snapped to the stream mouth (i.e., no negative distances)

```{r load_files}
# Import and make sure it's projected into AKAlbers
rivernetwork <- line2network(path = shapefile_path, layer = shapefile_name, reproject = AKalbers)
```

```{r plotnetwork}
# Show the river network and all segments, this should match the ArcMap shapefiles.
plot(x= rivernetwork)
```

NOTE - Run this chunk in console as the notebook lags 1 graph behind, making it difficult to answer questions...

    Cleanup steps taken for <i>STOCKDALE CREEK</i>:

    Insert verticies to increase snapping precision : y
    Minimum distance to use:  1 [this is in map units shown under the "load_files" chunk]
    Identify segment number of river mouth:  2
    Identify vertex number of river mouth:  1122
    Accept mouth assignment: y
    Remove additional segments:  n
    Build segment routes:  y

    Cleanup steps taken for <i>HOGAN CREEK</i>:

    Insert verticies to increase snapping precision : y
    Minimum distance to use:  1 [this is in map units shown under the "load_files" chunk]
    Identify segment number of river mouth:  1
    Identify vertex number of river mouth:  683
    Accept mouth assignment: y
    Remove additional segments:  n
    Build segment routes:  y

```{r CleanNetwork}
# run in console if you want / need visual guides
rivercleaned <- riverdist::cleanup(rivernetwork)

# check topology so network distance can be calculated, make sure all stream confluences are green and stream sources/outlets are red
# Run this twice for a plot...
topologydots(rivers = rivercleaned)
```

Import the raw field data with GPS locations and check for common errors, including missing Lat/Long (i.e., 0s) and instances where Longitude is missing the negative. Note that missing negatives could be a sign of larger data issues. It's probably best to toss it into ArcMap or equivalent and verify the locations prior to performing the blanket mutate below.

```{r import_data}
# Import the entire dataset and filter for selected stream
raw_dat <- read_csv(file = stream_data) %>% 
  filter( str_detect( StreamName, regex(stream, ignore_case = TRUE))) %>% # Overcomplicated, but needs the regex call to include ignore case since data has caps and GIS does not
  filter(Latitude != 0 | Longitude != 0) %>%  # toss if missing GPS info
  mutate(Longitude = case_when( Longitude > 0 ~ (-1 * Longitude),
                                TRUE ~ Longitude)) # convert positive longitudes to negative

tide_dat <- read_csv(file = tide_data) %>% 
  filter( str_detect( Stream, regex(stream, ignore_case = TRUE))) # Overcomplicated, but needs the regex call to include ignore case since data has caps and GIS does not
```

After importing the data, the coordinates must be converted into AKalberts xy format. This makes sure these are the same projection as the stream.

```{r xy2akalbers}
# pull xy coordinates and put into riverdist format
xy_fish <- cbind(as.numeric(raw_dat$Longitude),
                 as.numeric(raw_dat$Latitude))

# Convert from Lat/Long dec. degree coordinates to the xy system.
# Note coords are long-lat, not lat-long.
xyfish2albers <- rgdal::project(xy_fish, proj=AKalbers)

# Make sure this is just the xy-coords
head(xyfish2albers)

# Add xy-coords back to original fish data (i.e., sample ID)
fish_locs <- raw_dat %>% 
  select(SampleTrayId, SampleCell, Latitude, Longitude) %>% 
  unite_("Sample", c("SampleTrayId", "SampleCell")) %>% 
  add_column(x = xyfish2albers[,1], y = xyfish2albers[,2])

head(fish_locs)

# pull xy coordinate for tides and put into riverdist format
xy_tide <- cbind(as.numeric(tide_dat$Longitude),
                  as.numeric(tide_dat$Latitude))

# Convert from Lat/Long dec. degree coordinates to the xy system.
# Note coords are long-lat, not lat-long.
xytide2albers <- rgdal::project(xy_tide, proj=AKalbers)

# Make sure this is just the xy-coords
head(xytide2albers)

# Add xy-coords back to original tide data (i.e., sample ID)
tide_locs <- tide_dat %>% 
  select(Stream) %>% 
  add_column(x = xytide2albers[,1], y = xytide2albers[,2])

head(tide_locs)
```

Once the coordinates are in the correct format, the fish can be snapped to the stream network. The tide marks have already been snapped in ArcMap, but can be done in the same manner as the fish, if necessary.

```{r CalculateSnapDistance}
# Snap xy-coords to the nearest streamline. 
fish_locs_snapped <- xy2segvert(x = fish_locs$x, y = fish_locs$y, rivers = rivercleaned)

# Add fishID (Sample = TrayID_Cell)
fish_locs_snapped <- fish_locs_snapped %>% 
  add_column(Sample = fish_locs$Sample,
             Latitude = fish_locs$Latitude,
             Longitude = fish_locs$Longitude)

# Should contain segment, vertex, snapping distance, and sample id
head(fish_locs_snapped)

# Visual of how far the waypoints were from the stream network (i.e., snapping distance)
# Limiting to 400m from streamline, could be more (likely incorrect) GPS points outside of this
ggplot(fish_locs_snapped, aes(x = snapdist)) +
  geom_histogram(binwidth = 10) +
  theme_bw()
  ggtitle(label = "Distance of GPS waypoint from river (m)")


# Snap xy-coords to the nearest streamline. 
tide_locs_snapped <- xy2segvert(x = tide_locs$x, y = tide_locs$y, rivers = rivercleaned)

# Add tideID (Sample = TrayID_Cell)
tide_locs_snapped <- tide_locs_snapped %>% 
  add_column(Stream = tide_locs$Stream)

# Should contain segment, vertex, snapping distance, and sample id
head(tide_locs_snapped)
```

View the fish waypoints on the stream network and compare snapped (blue) to raw (red) locations:

```{r RawVsSnapped}
# Plot the streamlines, adding 100m buffer in ea direction so we can see all datapoints
plot(x = rivercleaned,
     xlim = c(min(fish_locs$x - 100), max(fish_locs$x + 100)),
     ylim = c(min(fish_locs$y - 100), max(fish_locs$y + 100)),
     xlab = "Longitude",
     ylab = "Latitude"
)

# Add raw data in red
points( x = fish_locs$x, y = fish_locs$y, pch =5, col = 2) 

# Add snapped data in blue
riverdist::riverpoints( seg = fish_locs_snapped$seg, vert=fish_locs_snapped$vert, rivers = rivercleaned, pch=10, col=4)

# Add legend
legend("topright", legend=c("raw", "snapped"),
       col=c("red", "blue"), lty=c(1,1), cex=0.8,
       box.lty=0)
```

Zoom in to get a look at the snapped high-tide marks

```{r intertidal_pts}
plot( x = rivercleaned)
points(tide_locs$x, tide_locs$y, pch=16, col="red")
```

Calculate the distance of every fish and tide mark to the river mouth.

```{r CalculateDistance}
# Calculate distance, using default algorithm. See riverdist docs for details of available algorithms.
mouthdist <- riverdist::mouthdist(seg = fish_locs_snapped$seg, vert = fish_locs_snapped$vert, rivers = rivercleaned)

# Calculate distance, using default algorithm. See riverdist docs for details of available algorithms.
tidedist <- riverdist::mouthdist(seg = tide_locs_snapped$seg, vert = tide_locs_snapped$vert, rivers = rivercleaned)
```

Create dataframe containing fish information and distance calculation. Also, add tide marks and distance information.

```{r FinalData}
fish_df <- fish_locs_snapped %>%  
  add_column(mouthdist)

# Should have segment, vertex, snapping distance, sample, and distance from mouth (start of streamline, not necessarily above tidal zone)
head(fish_df)


tide_df <- tide_locs_snapped %>%  
  add_column(tidedist)

# Should have segment, vertex, snapping distance, sample, and distance from mouth (start of streamline, not necessarily above tidal zone)
head(tide_df)
```

Plot of distance data by Lat/Long - These should be a continuous color scale

```{r QC_fishpts}
# Need to add xy coords to the fish_df for QC
# Again, limiting to only points within 400m of stream for visual
joined <- left_join(x = fish_df, y = fish_locs)


# Plot of rivernetwork which should be a continuous scale going upstream (points further from mouth are lighter colors)
ggplot(joined, aes(x = x, y= y, color = mouthdist)) + 
  geom_point() +
  coord_equal() +
  theme_bw() +
  ggtitle("Riverdist calculations")
```

# Tidal calulations

Now that distance to high-tide and distance to each fish have been calculated, we can calculate the distance above high-tide mark. Distances below the tide mark will be negative and distances above positive. Note that distances below tidal marks will still have positive river distances (distance from mouth), because the mouth does not equal high-tide mark. Generally, the 'mouth' is poorly defined for these streams. Further, since river polylines were extended or manually created to accomodate fish GPS points, the distance above/below high-tide point may be a more informative measure.

In some cases, there are two high-tide marks, one for each channel or tributary. Distances have to be manually assigned based on which tide mark is applicable to each fish.

```{r upstream_tidal}
tide_df$tidedist # view dataframe - are there multiple points?

final_df <- fish_df %>% 
  add_column(hitide = tide_df$tidedist) %>% 
  mutate(dist2tide = mouthdist - hitide)

final_df
```

```{r tidal_plot}
sum(final_df$dist2tide > 1) / nrow(final_df) # portion of waypoints are above the intertidal area


ggplot(final_df, aes(x = dist2tide)) +
  geom_histogram() +
  geom_vline(xintercept = 0) +
  theme_bw() +
  ggtitle(label = "Distance of GPS waypoint from intertidal mark (m)")
```

# Export Data

If all looks good, write a csv containing:

    Sample 
    Latitude
    Longitude
    Distance from river mouth
    Segment wpt was snapped to
    Vertex wpt was snapped to
    Snapping distance (how far wpt was from stream)
    Intertidal or not (dist2tide = negative/positive)

```{r exportCSV}
# Output csv
write_csv(x = final_df, file = output_path)
```

# Join with Parentage data

```{r get_riverdist}
(riverdist <- read_csv(output_path) %>% 
   mutate(Intertidal = case_when(dist2tide > 0 ~ "Upstream",
                                 dist2tide <= 0 ~ "Intertidal")) %>% 
   rename(Distance = mouthdist, Segment = seg))
```

# Join *riverdist* table with parentage data

We joined spatial data from *riverdist* with parentage data.

```{r join tables}
paired_14_16_filter_parents <- read_csv(file = paste0("../data/", stream, "_14_16/", stream, "_paired_14_16_filter_parents.csv"))

(
  riverdist_parentage <- paired_14_16_filter_parents %>% 
    unite("Sample", c("DNA Tray Code", "DNA Tray Well Code"), sep = "_", remove = FALSE) %>% 
    left_join(riverdist, by = "Sample") %>% 
    mutate(origin = factor(origin, c("Natural", "Hatchery")))  %>% 
    mutate(`Sample Date` = lubridate::ymd(`Sample Date`)) %>% 
    filter(is.na(`Length Mm`) | `Length Mm` > 300) %>% 
    mutate(Segment = factor(Segment))%>% 
    mutate(Success = case_when(n == 0 ~ "Unsuccessful",
                               n > 0 ~ "Successful"))
)
```

Write it out

```{r}
write_csv(x = riverdist_parentage, file = paste0("../data/", stream, "_14_16/riverdist_parentage_", stream, ".csv"))
```
